Fix permissions not updated in HPB when general permissions change
When the room or call permissions are set the permissions of all
participants could be modified. However, in this case the signaling
permissions were not updated, so the signaling server could allow or
reject media from participants that were already in the room when the
permissions changed. Now a message to update all the participants
currently in the room is sent to the signaling server when the room or
call permissions are set.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>