Revert "Fix participant permissions not reloaded when general permissions change"
This reverts commit 99def03db2d5470e4aaf139152c01bf79fc01c48.

The previous fix was flawed, as the participant permissions were
reloaded only for the participant who changed the general permissions.
Moreover, now all the participants reload the participant list both with
and without HPB when the general permissions change, so the previous fix
is no longer needed even for the participant who made the change.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>